♻️ refactor: #25 예약 확인 코드 정리
